Replace fish wrapper for nvm with fish-nvm
fish-nvm is native, does not require bass. Remove bass.

Remove fzf bin from PATH setting in config.fish. I use fzf vim plugin,
and it is setup to launch fzf installation script on :PlugUpd. This
script does a lot of things, once of which is add fzf bin to
$fish_user_paths. Thus, fzf bin was duplicated in $PATH because I was
also appending it in config.fish.